cortex-m: Add initial dwt and dcb support
The DWT and DCB peripherals are used to access various debug functionality
in the chip, such as various counters.
This commits adds register descriptions for most of DWT and all of the DCB.
It also adds a very simple capsule that allows starting, stopping, reading,
and resetting the cycle counter from userspace.